+++ ng-workspace/projects/OpenWaterFoundation/common/leaflet/map/map.module.ts
@@ -16,6 +16,9 @@ import { LegendLayerGroupModule }      from './legend/legend-layer-group/legend-
   declarations: [
     MapComponent
   ],
+  exports: [
+    MapComponent
+  ],
   imports: [
     CommonModule,
     BrowserAnimationsModule,