+++ ng-workspace/projects/OpenWaterFoundation/common/leaflet/public-api.ts
@@ -1,5 +1,7 @@
 export * from './map/map.component';
 export * from './map/map.module';
+export * from './map/map.util';
+
 export * from './map/legend/legend-background-group/legend-background-group.component';
 export * from './map/legend/legend-background-group/legend-background-group.module';
 export * from './map/legend/legend-layer-group/legend-layer-group.component';