+++ ng-workspace/projects/OpenWaterFoundation/common/src/public-api.ts
@@ -12,8 +12,10 @@ export * from '@OpenWaterFoundation/common/ts';
 export * from '@OpenWaterFoundation/common/ts-command-processor/commands/delimited';
 export * from '@OpenWaterFoundation/common/ts-command-processor/core';
 
+export * from '@OpenWaterFoundation/common/ui/dashboard';
 export * from '@OpenWaterFoundation/common/ui/dialog';
 export * from '@OpenWaterFoundation/common/ui/layer-manager';
+export * from '@OpenWaterFoundation/common/ui/widget';
 export * from '@OpenWaterFoundation/common/ui/window-manager';
 
 export * from '@OpenWaterFoundation/common/util/io';