+++ ng-workspace/projects/OpenWaterFoundation/common/ui/widget/image/image.module.ts
@@ -1,0 +1,21 @@
+import { NgModule }       from '@angular/core';
+import { CommonModule }   from '@angular/common';
+
+import { ErrorModule } from '../error/error.module';
+import { ImageComponent } from './image.component';
+
+
+@NgModule({
+  declarations: [
+    ImageComponent
+  ],
+  exports: [
+    ImageComponent
+  ],
+  imports: [
+    CommonModule,
+
+    ErrorModule
+  ]
+})
+export class ImageModule { }