+++ ng-workspace/projects/angulardev/src/app/app-routing.module.ts
@@ -3,6 +3,7 @@ import { RouterModule,
           Routes}             from '@angular/router';
           
 import { MapComponent }       from '@OpenWaterFoundation/common/leaflet';
+import { DashboardComponent } from '@OpenWaterFoundation/common/ui/dashboard';
 import { HomeComponent }      from './home/home.component';
 // import { MapErrorComponent }   from './map-components/map-error/map-error.component';
 import { NotFoundComponent }  from './not-found/not-found.component';
@@ -13,6 +14,7 @@ const routes: Routes = [
   { path: 'owf-common', component: OwfCommonComponent },
   { path: 'home', component: HomeComponent },
   { path: 'map/:id', component: MapComponent },
+  { path: 'dashboard', component: DashboardComponent },
   // { path: 'map-error', component: MapErrorComponent },
   { path: '404', component: NotFoundComponent },
   { path: '**', component: NotFoundComponent }